c315: extract BenchmarksDeep tab cluster (DeepKCurveSection + Cae3dAnchorVsFullSection + BetaVaeCollapseSection + AnomalyComparisonSection) to own file (#441 P0 2.2)
Benchmarks.tsx 2120 LOC -> 1620 LOC (-500 LOC). Benchmarks chunk 49.11 KB -> 35.83 KB (-13.3 KB).
New chunk: BenchmarksDeep-BNst8wK4.js 13.82 KB.
typecheck clean, 44/44 tests pass.
